+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed missed render on `unloadData` event when `modelfree` extension is used
 - Added visual guide lines for expanded objects in `signature` view, similar to `struct` view
 - Fixed rendering of copy source button in `source` view when a string is passed as data
+- Added propagation of meaningful `source` view props into nested views rendering (action buttons, prelude, and postlude) through the context as `sourceViewProps`
 
 ## 1.0.0-beta.91 (24-11-2024)